--- src/app/cancel/page.tsx
@@ -1,11 +1,0 @@
-const CancelPage = () => {
-    return (
-      <div>
-        <h1>Payment Canceled</h1>
-        <p>Your payment has been canceled.</p>
-      </div>
-    );
-  };
-
-  export default CancelPage;
-
