+++ src/app/layout.tsx
@@ -4,7 +4,7 @@ import "../styles/globals.css";
 import Header from "@/components/Header/Header";
 import Sidebar from "@/components/Sidebar";
 import Footer from "@/components/Footer";
-import {FormProvider} from '@/components/FormContext'
+import { FormProvider } from "@/components/FormContext";
 import FooterDesctop from "@/components/FooterDesctop";
 
 const lato = Lato({
@@ -17,7 +17,6 @@ export const metadata: Metadata = {
   description: "Cleaning service Atlanta",
 };
 
-
 export default function RootLayout({
   children,
 }: Readonly<{
@@ -27,11 +26,11 @@ export default function RootLayout({
     <html lang="en">
       <body className={lato.className}>
         <FormProvider>
-        <Header />
-        <Sidebar />
-        <main className="lg:ml-[200px] xl:ml-[244px]">{children}</main>
-        <Footer />
-        <FooterDesctop />
+          <Header />
+          <Sidebar />
+          <main className="lg:ml-[200px] xl:ml-[244px]">{children}</main>
+          <Footer />
+          <FooterDesctop />
         </FormProvider>
       </body>
     </html>
